pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "molid"
version = "0.2.0"
description = "A package for processing PubChem data and managing a database of chemical compounds."
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "Thea Denell", email = "thea.denell@physik.hu-berlin.de"}
]
license = {file = "LICENSE"}
classifiers = [
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Physics",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Operating System :: OS Independent"
]

dependencies = [
    "ase>=1.1.7",
    "psutil == 6.1.1",
    "pytest",
    "openbabel-wheel"
]

[project.scripts]
molid-query      = "molid.cli:main"
molid-db-manager = "molid.db.db_manager:main"

[tool.setuptools.packages.find]
include = ["molid*"]  # Ensure only the 'molid' package is included
exclude = ["processed", "downloads", "tests"]  # Exclude non-package directories
